Release 3.2.0

+++ l4/l4_e2/src/main.c
@@ -42,7 +42,7 @@ static K_SEM_DEFINE(run_app, 0, 1);
 static uint8_t client_id[sizeof(CONFIG_BOARD) + 11];
 
 
-static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				   struct net_if *iface)
 {
 	if ((mgmt_event & EVENT_MASK) != mgmt_event) {
